Added a cast in vector2d to make MSVC happy with integer vectors. Added virtual destructors in factory interfaces.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@1002 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/vector2d.h
@@ -67,7 +67,7 @@ public:
 
 	//! Returns the length of the vector
 	//! \return Returns the length of the vector.
-	T getLength() const { return (T)sqrt(X*X + Y*Y); }
+	T getLength() const { return (T)sqrt((f64)(X*X + Y*Y)); }
 
 	//! Returns the squared length of this vector
 	/** This is useful because it is much faster than getLength(). */
